refactor(group): createFreeGroupPlugin move to container-plugin
