+++ apps/demo-free-layout/src/components/group/components/background.tsx
