+++ apps/demo-free-layout/src/components/group/components/tips/global-store.ts
