+++ apps/demo-free-layout/src/components/group/components/tips/is-mac-os.ts
