+++ apps/demo-free-layout/src/hooks/use-editor-props.tsx
@@ -7,18 +7,20 @@ import { createFreeSnapPlugin } from '@flowgram.ai/free-snap-plugin';
 import { createFreeNodePanelPlugin } from '@flowgram.ai/free-node-panel-plugin';
 import { createFreeLinesPlugin } from '@flowgram.ai/free-lines-plugin';
 import { FreeLayoutProps } from '@flowgram.ai/free-layout-editor';
-import { createContainerNodePlugin } from '@flowgram.ai/free-container-plugin';
+import {
+  createContainerNodePlugin,
+  createFreeGroupPlugin,
+} from '@flowgram.ai/free-container-plugin';
 
 import { onDragLineEnd } from '../utils';
 import { FlowNodeRegistry, FlowDocumentJSON } from '../typings';
 import { shortcuts } from '../shortcuts';
 import { CustomService } from '../services';
-import { createFreeGroupPlugin } from '../plugins/free-group-plugin';
 import { createSyncVariablePlugin } from '../plugins';
 import { defaultFormMeta } from '../nodes/default-form-meta';
 import { WorkflowNodeType } from '../nodes';
 import { SelectorBoxPopover } from '../components/selector-box-popover';
-import { BaseNode, CommentRender, LineAddButton, NodePanel } from '../components';
+import { BaseNode, CommentRender, GroupNodeRender, LineAddButton, NodePanel } from '../components';
 
 export function useEditorProps(
   initialData: FlowDocumentJSON,
@@ -226,7 +228,9 @@ export function useEditorProps(
          * 这个用于 loop 节点子画布的渲染
          */
         createContainerNodePlugin({}),
-        createFreeGroupPlugin({}),
+        createFreeGroupPlugin({
+          groupNodeRender: GroupNodeRender,
+        }),
       ],
     }),
     []
