+++ packages/plugins/free-container-plugin/package.json
@@ -1,66 +1,67 @@
 {
-  "name": "@flowgram.ai/free-container-plugin",
-  "version": "0.1.8",
-  "homepage": "https://flowgram.ai/",
-  "repository": "https://github.com/bytedance/flowgram.ai",
-  "license": "MIT",
-  "exports": {
+    "name": "@flowgram.ai/free-container-plugin",
+    "version": "0.1.8",
+    "homepage": "https://flowgram.ai/",
+    "repository": "https://github.com/bytedance/flowgram.ai",
+    "license": "MIT",
+    "exports": {
+        "types": "./dist/index.d.ts",
+        "import": "./dist/esm/index.js",
+        "require": "./dist/index.js"
+    },
+    "main": "./dist/index.js",
+    "module": "./dist/esm/index.js",
     "types": "./dist/index.d.ts",
-    "import": "./dist/esm/index.js",
-    "require": "./dist/index.js"
-  },
-  "main": "./dist/index.js",
-  "module": "./dist/esm/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "npm run build:fast -- --dts-resolve",
-    "build:fast": "tsup src/index.ts --format cjs,esm --sourcemap --legacy-output",
-    "build:watch": "npm run build:fast -- --dts-resolve",
-    "clean": "rimraf dist",
-    "test": "exit 0",
-    "test:cov": "exit 0",
-    "ts-check": "tsc --noEmit",
-    "watch": "npm run build:fast -- --dts-resolve --watch --ignore-watch dist"
-  },
-  "dependencies": {
-    "@flowgram.ai/free-history-plugin": "workspace:*",
-    "@flowgram.ai/free-lines-plugin": "workspace:*",
-    "@flowgram.ai/core": "workspace:*",
-    "@flowgram.ai/document": "workspace:*",
-    "@flowgram.ai/free-layout-core": "workspace:*",
-    "@flowgram.ai/renderer": "workspace:*",
-    "@flowgram.ai/utils": "workspace:*",
-    "inversify": "^6.0.1",
-    "reflect-metadata": "~0.2.2",
-    "lodash": "^4.17.21"
-  },
-  "devDependencies": {
-    "@flowgram.ai/eslint-config": "workspace:*",
-    "@flowgram.ai/ts-config": "workspace:*",
-    "@types/bezier-js": "4.1.3",
-    "@types/lodash": "^4.14.137",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "@types/styled-components": "^5",
-    "@vitest/coverage-v8": "^0.32.0",
-    "eslint": "^8.54.0",
-    "react": "^18",
-    "react-dom": "^18",
-    "styled-components": "^5",
-    "tsup": "^8.0.1",
-    "typescript": "^5.0.4",
-    "vitest": "^0.34.6"
-  },
-  "peerDependencies": {
-    "react": ">=17",
-    "react-dom": ">=17",
-    "styled-components": ">=4"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  }
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "build": "npm run build:fast -- --dts-resolve",
+        "build:fast": "tsup src/index.ts --format cjs,esm --sourcemap --legacy-output",
+        "build:watch": "npm run build:fast -- --dts-resolve",
+        "clean": "rimraf dist",
+        "test": "exit 0",
+        "test:cov": "exit 0",
+        "ts-check": "tsc --noEmit",
+        "watch": "npm run build:fast -- --dts-resolve --watch --ignore-watch dist"
+    },
+    "dependencies": {
+        "@flowgram.ai/free-history-plugin": "workspace:*",
+        "@flowgram.ai/free-lines-plugin": "workspace:*",
+        "@flowgram.ai/core": "workspace:*",
+        "@flowgram.ai/document": "workspace:*",
+        "@flowgram.ai/shortcuts-plugin": "workspace:*",
+        "@flowgram.ai/free-layout-core": "workspace:*",
+        "@flowgram.ai/renderer": "workspace:*",
+        "@flowgram.ai/utils": "workspace:*",
+        "inversify": "^6.0.1",
+        "reflect-metadata": "~0.2.2",
+        "lodash": "^4.17.21"
+    },
+    "devDependencies": {
+        "@flowgram.ai/eslint-config": "workspace:*",
+        "@flowgram.ai/ts-config": "workspace:*",
+        "@types/bezier-js": "4.1.3",
+        "@types/lodash": "^4.14.137",
+        "@types/react": "^18",
+        "@types/react-dom": "^18",
+        "@types/styled-components": "^5",
+        "@vitest/coverage-v8": "^0.32.0",
+        "eslint": "^8.54.0",
+        "react": "^18",
+        "react-dom": "^18",
+        "styled-components": "^5",
+        "tsup": "^8.0.1",
+        "typescript": "^5.0.4",
+        "vitest": "^0.34.6"
+    },
+    "peerDependencies": {
+        "react": ">=17",
+        "react-dom": ">=17",
+        "styled-components": ">=4"
+    },
+    "publishConfig": {
+        "access": "public",
+        "registry": "https://registry.npmjs.org/"
+    }
 }
