+++ packages/plugins/free-container-plugin/src/group/shortcuts/group.ts
@@ -1,8 +1,6 @@
-import {
-  FreeLayoutPluginContext,
-  ShortcutsHandler,
-  WorkflowSelectService,
-} from '@flowgram.ai/free-layout-editor';
+import { ShortcutsHandler } from '@flowgram.ai/shortcuts-plugin';
+import { WorkflowSelectService } from '@flowgram.ai/free-layout-core';
+import { PluginContext } from '@flowgram.ai/core';
 
 import { WorkflowGroupService } from '../workflow-group-service';
 import { WorkflowGroupCommand } from '../constant';
@@ -20,7 +18,7 @@ export class GroupShortcut implements ShortcutsHandler {
 
   private groupService: WorkflowGroupService;
 
-  constructor(context: FreeLayoutPluginContext) {
+  constructor(context: PluginContext) {
     this.selectService = context.get(WorkflowSelectService);
     this.groupService = context.get(WorkflowGroupService);
     this.execute = this.execute.bind(this);
