+++ packages/plugins/free-container-plugin/src/group/shortcuts/index.ts
