+++ packages/plugins/free-container-plugin/src/group/type.ts
@@ -1,0 +1,6 @@
+import { FC } from 'react';
+
+export interface WorkflowGroupPluginOptions {
+  groupNodeRender: FC;
+  disableGroupShortcuts?: boolean;
+}
