+++ packages/plugins/free-container-plugin/src/index.ts
@@ -1,4 +1,3 @@
 export * from './node-into-container';
 export * from './sub-canvas';
-export { createContainerNodePlugin } from './plugin';
-export type { WorkflowContainerPluginOptions } from './type';
+export * from './group';
