+++ packages/plugins/free-container-plugin/src/node-into-container/index.ts
@@ -1,3 +1,8 @@
 export { NodeIntoContainerType } from './constant';
 export { NodeIntoContainerService } from './service';
-export { NodeIntoContainerState, NodeIntoContainerEvent } from './type';
+export {
+  NodeIntoContainerState,
+  NodeIntoContainerEvent,
+  WorkflowContainerPluginOptions,
+} from './type';
+export { createContainerNodePlugin } from './plugin';
