+++ packages/plugins/free-container-plugin/src/node-into-container/type.ts
@@ -18,3 +18,7 @@ export interface NodeIntoContainerEvent {
   sourceContainer?: WorkflowNodeEntity;
   targetContainer: WorkflowNodeEntity;
 }
+
+export interface WorkflowContainerPluginOptions {
+  disableNodeIntoContainer?: boolean;
+}
