--- packages/plugins/free-container-plugin/src/type.ts
@@ -1,3 +1,0 @@
-export interface WorkflowContainerPluginOptions {
-  disableNodeIntoContainer?: boolean;
-}
